+++ src/components/Forms/NewsletterForm/useNewsletterForm.ts
@@ -3,16 +3,28 @@ import { useCallback } from "react";
 import { hubspotSubmitForm } from "../../../browser-lib/hubspot/forms";
 import { HubspotFormData } from "../../../browser-lib/hubspot/forms/hubspotSubmitForm";
 import config from "../../../config";
+import useAnalytics from "../../../context/Analytics/useAnalytics";
+import useAnonymousId from "../../../browser-lib/analytics/useAnonymousId";
 
 const hubspotNewsletterFormId = config.get("hubspotNewsletterFormId");
 
 const useNewsletterForm = () => {
-  const onSubmit = useCallback((data: HubspotFormData) => {
-    return hubspotSubmitForm({
-      hubspotFormId: hubspotNewsletterFormId,
-      data,
-    });
-  }, []);
+  const anonymousId = useAnonymousId();
+
+  const { identify } = useAnalytics();
+  const onSubmit = useCallback(
+    (data: HubspotFormData) => {
+      const hubspotFormResponse = hubspotSubmitForm({
+        hubspotFormId: hubspotNewsletterFormId,
+        data,
+      });
+
+      identify(anonymousId, { email: data.email || data.emailTextOnly });
+
+      return hubspotFormResponse;
+    },
+    [identify, anonymousId]
+  );
   return { onSubmit };
 };